Reorganize and optimize.
Move the server-side code up a directory level to the root.
Rename the client-side "src" directory to "client".

Move zxcvbn to a separate webpack bundle, which will be included
only when visiting a page with a password entry form.
Switch from moment to moment-mini (without locale support).
Load specific bootstrap-vue components instead of the whole suite.