Change email header order.
By spec, the order of headers in an email isn't supposed to matter.
Empirically, however, some SMTP servers reject (as spam) mail whose
headers aren't in an order they like.  There doesn't seem to be any
rhyme or reason to it (or any documentation of it), but this order seems
to work.

Also removed debugging from previous commits.