Switch unsubscribe page to use email.
Mailman doesn't support putting an opaque unsubscribe token in email
footers, so the unsubscribe link now takes the caller's email address.
This could allow someone to unsubscribe someone else, but hopefully we'd
catch that reviewing the daily logs.

Note that the code still populates and maintains an unsubscribe token in
each person's record, because getting rid of it would mean changing the
database schema, which is a pain.  However, this token is no longer
used.